requirements.txt
numpy
pandas
scikit-learn
pytorch-forecasting
pytorch-lightning
fastapi
uvicorn
python-multipart
networkx
xgboost
lightgbm
matplotlib
seaborn
requests
